wip: project file serialization/deserialization, logging of problems

+++ Laura/src/Project/Scene/SceneManager.cpp
@@ -12,15 +12,20 @@ namespace Laura
     void SceneManager::RemoveScene(LR_GUID guid) {
         if (m_ActiveSceneGUID == guid)  { m_ActiveSceneGUID = LR_GUID::INVALID; }
         if (m_BootSceneGUID == guid)    { m_BootSceneGUID   = LR_GUID::INVALID; }
-        m_Scenes.erase(guid);
+        size_t removed = m_Scenes.erase(guid);
+        if (removed == 0) {
+            LOG_ENGINE_WARN("Scene to be erased Not Found");
+        }
     }
 
     bool SceneManager::SetActiveScene(LR_GUID guid) {
         if (guid == LR_GUID::INVALID) {
+            LOG_ENGINE_WARN("Invalid GUID passed")
             return false;
         }
         auto it = m_Scenes.find(guid); 
         if (it == m_Scenes.end()) {
+            LOG_ENGINE_WARN("Scene Not found");
             return false;
         }
         m_ActiveSceneGUID = guid; 
@@ -29,11 +34,13 @@ namespace Laura
 
     std::shared_ptr<Scene> SceneManager::GetActiveScene() const {
         if (m_ActiveSceneGUID == LR_GUID::INVALID){
+            LOG_ENGINE_WARN("Invalid GUID passed")
             return nullptr;
         }
         if (auto it = m_Scenes.find(m_ActiveSceneGUID); it != m_Scenes.end()) {
             return it->second;
         }
+        LOG_ENGINE_WARN("Scene Not found");
         return nullptr;
     }
 
@@ -43,12 +50,17 @@ namespace Laura
 		// Remove orphaned scene files
 		for (const auto& filepath : FindFilesInFolder(projectRoot, SCENE_FILE_EXTENSION)) {
 			LR_GUID guid = GuidFromFilename(filepath.filename());
+            if (guid == LR_GUID::INVALID) {
+                LOG_ENGINE_WARN("Unable to delete scene file: {0}", filepath);
+                continue;
+            }
 
 			if (m_Scenes.find(guid) == m_Scenes.end()) {
 				std::error_code ec;
 				std::filesystem::remove(filepath, ec);
 				if (ec) {
 					successAll = false;
+                    LOG_ENGINE_WARN("Unable to delete scene file: {0}", filepath);
 				}
 			}
 		}
@@ -58,6 +70,7 @@ namespace Laura
             auto filepath = SceneFilepathFromGuid(projectRoot, scene->GetGUID());
 			if (!scene->Serialize(filepath)) {
 				successAll = false;
+                LOG_ENGINE_WARN("Unable to serialize scene file: {0}", filepath);
 			}
 		}
 
@@ -71,6 +84,7 @@ namespace Laura
             bool success = scene->Deserialize(scenepath);
             if (!success) { 
                 successAll = false; 
+                LOG_ENGINE_WARN("Unable to Deserialize scene file: {0}", filepath);
                 continue;
             }
             m_Scenes[scene->GetGUID()] = scene;
@@ -81,9 +95,9 @@ namespace Laura
     static std::vector<std::filesystem::path> SceneManager::FindFilesInFolder(
             const std::filesystem::path& folder, 
             const std::string& extension) {
-
         std::vector<std::filesystem::path> result;
 		if (!std::filesystem::exists(folder) || !std::filesystem::is_directory(folder)) {
+            LOG_ENGINE_WARN("Folder does not exists or is not a directory {0}", folder);
 			return result;
 		}
 		for (const auto& entry : std::filesystem::directory_iterator(folder)) {
@@ -105,6 +119,7 @@ namespace Laura
             return LR_GUID(value);
         }
         catch (const std::exception&) {
+            LOG_ENGINE_WARN("Unable to extract GUID from filename: {0}", filepath);
             return LR_GUID::INVALID;
         }
     }
